backend/pyproject.toml
[project]
name = "posiedon-backend"
version = "0.1.0"
description = "Posiedon - ç®€æ´é€šç”¨çš„ç¾¤ä½“æ™ºèƒ½å¼•æ“Žï¼Œé¢„æµ‹ä¸‡ç‰©"
requires-python = ">=3.11"
license = { text = "AGPL-3.0" }
authors = [
    { name = "Posiedon Team" }
]

dependencies = [
    # æ ¸å¿ƒæ¡†æž¶
    "flask>=3.0.0",
    "flask-cors>=6.0.0",
    
    # LLM ç›¸å…³
    "openai>=1.0.0",
    
    # Zep Cloud
    "zep-cloud==3.13.0",
    
    # OASIS ç¤¾äº¤åª’ä½“æ¨¡æ‹Ÿ
    "camel-oasis==0.2.5",
    "camel-ai==0.2.78",
    
    # æ–‡ä»¶å¤„ç†
    "PyMuPDF>=1.24.0",
    # ç¼–ç æ£€æµ‹ï¼ˆæ”¯æŒéžUTF-8ç¼–ç çš„æ–‡æœ¬æ–‡ä»¶ï¼‰
    "charset-normalizer>=3.0.0",
    "chardet>=5.0.0",
    
    # å·¥å…·åº“
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pipreqs>=0.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]
